requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asyncio==3.4.3
attrs==17.2.0
autobahn==17.9.2
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
Django==1.11.4
django-channels-panel==0.0.5
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-filter==1.0.4
django-multiselectfield==0.1.8
django-rest-auth==0.9.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
feedparser==5.2.1
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
PyJWT==1.5.2
pytz==2017.2
redis==2.10.6
six==1.10.0
sqlparse==0.2.3
Twisted==17.5.0
txaio==2.8.2
websocket-client==0.44.0
zope.interface==4.4.2